ABI: fix desynchronization of argument indices
When a shadow return value location argument is distributed, it is
shifting all the other argument indices by one. This was not taken into
the account previously. This commit fixes the indices.

It also makes it so the arguments `SizeOnStack` does not account for
padding. This is useful for some layout users.