Sugiyama: rework artifical entry computation
This refactors the way artifician entries are kept track of by allowing
an already existing entry to be updated with new successors.

It also introduces an improved version of artifical entry cleanup that
properly corrects the cases where the edge from an artificial entry
to a real one was a long edge, leading to it being split with
the leftovers never leaving the graph. This was leading to horizontal
lanes to appear where they shouldn't be - thus spacing nodes further
away from each other than necessary.

On top of that, a bunch of obsolete checks to ensure the graph is
"actually" single entry are removed. They are not necessary anymore,
with the `ensureSingleEntry` being more trustworthy, and so they are
just clogging debug builds for no reason.